requirements.txt
Flask==3.0.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
